Move things to activity.cpp
