+++ src/variables.hpp
@@ -1045,7 +1045,6 @@ void supply_income();
 void initialize_recipememory();
 void initialize_recipe();
 void label_1933();
-void label_19342();
 void label_1935();
 turn_result_t step_into_gate();
 void label_1955();
@@ -1065,7 +1064,6 @@ turn_result_t try_interact_with_npc();
 void sort_list_by_column1();
 void sort_list_and_listn_by_column1();
 void label_2076();
-void prompt_stop_continuous_action();
 void try_to_return();
 turn_result_t do_gatcha();
 void dump_player_info();
@@ -1088,9 +1086,6 @@ turn_result_t do_debug_console();
 turn_result_t do_exit_debug_console();
 void disarm_trap();
 void proc_trap();
-void continuous_action_perform();
-void continuous_action_sex();
-void continuous_action_others();
 void label_2149();
 void label_2150();
 void label_2151();
@@ -1187,9 +1182,7 @@ void removeitem(int, int);
 void removeworker(int = 0);
 void ride_begin(int = 0);
 void ride_end();
-void rowact_check(int = 0);
-void rowact_item(int = 0);
-void rowactend(int = 0);
+void rm_crlf(std::string&);
 void saccard(int = 0, int = 0);
 void savecycle();
 void setunid(int = 0, int = 0);
